refactor(activity): share the timeline merge between activity commands
`projects activity` was built to mirror `issues activity`, and its timeline
plumbing — `TimelineEntry`, `compareTimelineEntries`, `TimelinePage` and
`paginateTimeline` — was copied over byte for byte, differing only in the
item type. Two copies of the cursor arithmetic is one too many: a fix to the
`--after` handling in one would not reach the other, and the two commands are
documented as behaving identically.

The four declarations move to `src/common/timeline.ts`, generic over the item
type. Each service keeps only its own `toTimelineEntry`, which is genuinely
per-domain — the two item unions tag their comment threads differently.

Generalizing `toTimelineEntry` too (via an accessor callback) was considered
and dropped: it would trade a three-line ternary for indirection at every
call site without removing any duplicated logic.
